fix(operator-ui): update to latest (#15296)
Allow user to manually input ChainId and Account Address when there are no selections available.

When creating a chain config in operator ui, if the core node is not configured with APTOS RPC, user wont be able to create a chain config because there are nothing to select from in the ChainId and Account Address field in the chain config popup.

This is a new scenario for Keystone where a chain config needs to be created even if there are no existing connection to an RPC.

Operator UI PR: https://github.com/smartcontractkit/operator-ui/pull/95

JIRA: https://smartcontract-it.atlassian.net/browse/DPA-1306