Move txdb and dialects to chainlink-common/pkg/pg (#15064)
* Move txdb.go and dialects.go to chainlink-common

Along with that:
- Replace dialects imports with pgcommon
- Replace testutils.Skip* with tests.Skip*
- Call RegisterTxDb from NewConnection
- Run goimports (somehow there were a few files on develop which were improperly formatted?)

* Fix merge conflict with copied txdb driver

While rebasing, I found that CCIP has copied the old txdb driver from
chainlink into chainlink/integrations-tests. The problem with duplicated
code is it makes it more difficult to update. Little did they know, it
had already been updated even before they copied it. So this removes the
copy and imports the newer version from chainlink-common.

The motivation for copying rather than import appears to have come from
a lint rule that enforces not being able to import anything from an
external package with "internal" in the directory name. This still
requires that the NewSqlxDB be copied, but that's just a wrapper
around the pgcommon version which reads the db url from the chainlik
toml config, which can't be accessed in chainlink-common.

* chainlink-common/pkg/pg -> chainlink-common/pkg/sqlutil/pg

* Fix lint introduced by rebase

* pg.NewSqlxDB -> pg.NewTestDB

* Update chainlink-common ref

* Fix merge conflict